Ar_app/requirements.txt
Django==1.8
django-crispy-forms==1.5.2
django-parsley==0.6
psycopg2==2.6.1
